+++ example/lib/src/common/routes/route_state_mixin.dart
@@ -8,6 +8,8 @@ mixin RouteStateMixin<T extends StatefulWidget> on State<T> {
 
   late ElixirGuard guards;
 
+  late List<ElixirRouteDefinition> routes;
+
   @override
   void initState() {
     super.initState();
@@ -15,6 +17,27 @@ mixin RouteStateMixin<T extends StatefulWidget> on State<T> {
 
     guards = [
       (context, state) => state.length > 1 ? state : [const HomePage()],
+    ];
+    routes = [
+      ElixirRouteDefinition(name: 'home', builder: (_) => const HomePage()),
+      ElixirRouteDefinition(
+        name: 'settings',
+        builder:
+            (arguments) =>
+                SettingsPage(data: arguments['data'] as String? ?? ''),
+      ),
+      ElixirRouteDefinition(
+        name: 'profile',
+        builder: (_) => const ProfilePage(),
+      ),
+      ElixirRouteDefinition(
+        name: 'details',
+        builder:
+            (arguments) => DetailsPage(
+              userId: arguments['userId'] as String? ?? '',
+              note: arguments['note'] as String? ?? '',
+            ),
+      ),
     ];
   }
 }